backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
dj-rest-auth>=5.0
django-allauth>=0.63
drf-spectacular>=0.27
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3
Pillow>=10.2
django-filter>=24.2
gunicorn>=22.0
whitenoise>=6.6
celery>=5.3
redis>=5.0
django-storages[boto3]>=1.14
boto3>=1.34
